services/query_service/requirements.txt
FLASK
psycopg2-binary
redis
requests